==> src/2_Splitting.py <==
"""
Splitting the dataset into a training and testing set
Author: Bart Schelpe
Filename: 2_Splitting.py
Dataset: Avocado Prices
Link: https://www.kaggle.com/datasets/neuromusic/avocado-prices
Code based on: Python4AI PowerPoint presentations and documentation of varying Python packages
"""

# import packages
import pandas as pd

# read dataframe from pickle file
df = pd.read_pickle("../data/df.pickle")

# shuffle the dataset
df = df.sample(frac=1, random_state=69)

print(df.head())

# Split the shuffled DataFrame into training and testing sets
train_size = int(0.8 * len(df))  # 80% for training
train_df = df[:train_size]
test_df = df[train_size:]

# save both dataframes to new pickle files
train_df.to_pickle("../data/dfTrain.pickle")
test_df.to_pickle("../data/dfTest.pickle")
